% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_support.R
\name{holidays_to_json}
\alias{holidays_to_json}
\title{Convert streaming rate to JSON for upload}
\usage{
holidays_to_json(holidays, energy_code, energy_desc, prettify = FALSE)
}
\arguments{
\item{holidays}{data.frame or data.table like the ones created by the import_holidays function}

\item{energy_code}{atomic character two character MIDAS energy code for uploading company}

\item{energy_desc}{atomic character name of uploading energy company from Energy lookup table}

\item{prettify}{atomic logical TRUE to generate prettified JSON, FALSE by default}
}
\description{
Function to encode the data.table output of the TOU_to_streaming function
into JSON for upload to MIDAS.
}
