% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_support.R
\name{import_holidays}
\alias{import_holidays}
\title{Holiday CSV import}
\usage{
import_holidays(holiday_file)
}
\arguments{
\item{holiday_file}{atomic character with the path to a CSV file with holiday dates and information}
}
\description{
Function to import CSV files that contain holiday information. CSV must contain the two columns
"holiday_name" and "date". The date column must be formatted unambiguously, YYYY-MM-DD is best.
Holiday dates should be in local time. For example, New Year's Day 2023 is 2023-01-01.
}
