% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas.R
\name{MIDAS}
\alias{MIDAS}
\title{Register a username and password}
\description{
This class holds all the necessary information to maintain a connection to
the API. It automatically retrieves a log-in token when necessary.
}
\details{
MIDAS R6 class that connects to the MIDAS Rate API
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{midas_url}}{base URL for SGIP API}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MIDAS-new}{\code{MIDAS$new()}}
\item \href{#method-MIDAS-print}{\code{MIDAS$print()}}
\item \href{#method-MIDAS-register}{\code{MIDAS$register()}}
\item \href{#method-MIDAS-get_token}{\code{MIDAS$get_token()}}
\item \href{#method-MIDAS-value}{\code{MIDAS$value()}}
\item \href{#method-MIDAS-rins}{\code{MIDAS$rins()}}
\item \href{#method-MIDAS-value_history}{\code{MIDAS$value_history()}}
\item \href{#method-MIDAS-history_rins}{\code{MIDAS$history_rins()}}
\item \href{#method-MIDAS-lookups}{\code{MIDAS$lookups()}}
\item \href{#method-MIDAS-holiday}{\code{MIDAS$holiday()}}
\item \href{#method-MIDAS-upload_rate}{\code{MIDAS$upload_rate()}}
\item \href{#method-MIDAS-upload_holiday}{\code{MIDAS$upload_holiday()}}
\item \href{#method-MIDAS-clone}{\code{MIDAS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIDAS-new"></a>}}
\if{latex}{\out{\hypertarget{method-MIDAS-new}{}}}
\subsection{Method \code{new()}}{
Create a new MIDAS connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIDAS$new(
  username,
  password,
  email = NA_character_,
  fullname = NA_character_,
  organization = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{atomic character Username for login}

\item{\code{password}}{atomic character Password for login}

\item{\code{email}}{atomic character Email address for registration. Required if registering a new account}

\item{\code{fullname}}{atomic character Full name of user registering for MIDAS. Required if registering a new account}

\item{\code{organization}}{atomic character organization name for registration. Not required. Only used when registering a new account.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new \code{MIDAS} connection object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIDAS-print"></a>}}
\if{latex}{\out{\hypertarget{method-MIDAS-print}{}}}
\subsection{Method \code{print()}}{
Print some important characteristics of the MIDAS object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIDAS$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIDAS-register"></a>}}
\if{latex}{\out{\hypertarget{method-MIDAS-register}{}}}
\subsection{Method \code{register()}}{
This function registers a username and password.
The CEC requires that you be registered and respond to an authentication email
before obtaining an access token. You should only have to do this once.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIDAS$register(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters passed to curl}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIDAS-get_token"></a>}}
\if{latex}{\out{\hypertarget{method-MIDAS-get_token}{}}}
\subsection{Method \code{get_token()}}{
Get a token. This will usually be automatic when using other functions.
MIDAS tokens are valid for 10 minutes according to the CEC documentation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIDAS$get_token(new_token = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_token}}{logical Force requesting a new token, FALSE by default}

\item{\code{...}}{additional parameters passed to curl}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIDAS-value"></a>}}
\if{latex}{\out{\hypertarget{method-MIDAS-value}{}}}
\subsection{Method \code{value()}}{
Get real-time or current and forecast rate information.
Passing an empty RIN will return the XSD for formatting uploads.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIDAS$value(rin, query_type = "realtime", new_token = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rin}}{character Single RIN including dashes to request}

\item{\code{query_type}}{character One of "realtime" or "alldata"}

\item{\code{new_token}}{logical Force requesting a new token, FALSE by default}

\item{\code{...}}{additional parameters passed to curl}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIDAS-rins"></a>}}
\if{latex}{\out{\hypertarget{method-MIDAS-rins}{}}}
\subsection{Method \code{rins()}}{
Get list of available rates from MIDAS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIDAS$rins(signal_type = 1, new_token = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal_type}}{integer One of 0 = ALL signal types, 1 = Tariff signals,
2 = Greenhouse Gas signals, 3 = Flex Alert signals}

\item{\code{new_token}}{logical Force requesting a new token, FALSE by default}

\item{\code{...}}{additional parameters passed to curl}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIDAS-value_history"></a>}}
\if{latex}{\out{\hypertarget{method-MIDAS-value_history}{}}}
\subsection{Method \code{value_history()}}{
Get historical rate information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIDAS$value_history(rin, start_date, end_date, new_token = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rin}}{character Single RIN including dashes to request}

\item{\code{start_date}}{character or Date or IDate Atomic date for start of period of interest}

\item{\code{end_date}}{character or Date or IDate Atomic date for end of period of interest}

\item{\code{new_token}}{logical Force requesting a new token, FALSE by default}

\item{\code{...}}{additional parameters passed to curl}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIDAS-history_rins"></a>}}
\if{latex}{\out{\hypertarget{method-MIDAS-history_rins}{}}}
\subsection{Method \code{history_rins()}}{
Get list of rates with history available from MIDAS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIDAS$history_rins(distribution_code, energy_code, new_token = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distribution_code}}{atomic character Two letter code from the distribution table for the distribution company}

\item{\code{energy_code}}{atomic character Two letter code from the energy table for the energy company}

\item{\code{new_token}}{logical Force requesting a new token, FALSE by default}

\item{\code{...}}{additional parameters passed to curl}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIDAS-lookups"></a>}}
\if{latex}{\out{\hypertarget{method-MIDAS-lookups}{}}}
\subsection{Method \code{lookups()}}{
Get lookup table information. Possible lookup tables are currently
Country, Daytype, Distribution, Enduse, Energy,
Location, RateType, Sector, State, and Unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIDAS$lookups(table_name, new_token = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{atomic character lookup table name.}

\item{\code{new_token}}{logical Force requesting a new token, FALSE by default}

\item{\code{...}}{additional parameters passed to curl}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIDAS-holiday"></a>}}
\if{latex}{\out{\hypertarget{method-MIDAS-holiday}{}}}
\subsection{Method \code{holiday()}}{
Get holiday table information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIDAS$holiday(new_token = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_token}}{logical Force requesting a new token, FALSE by default}

\item{\code{...}}{additional parameters passed to curl}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIDAS-upload_rate"></a>}}
\if{latex}{\out{\hypertarget{method-MIDAS-upload_rate}{}}}
\subsection{Method \code{upload_rate()}}{
Requires special CEC approval. Upload access is
generally only granted to load serving entities, with rare exceptions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIDAS$upload_rate(filename, upload_format = "xml", verbose = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{path to file with XML to upload}

\item{\code{upload_format}}{character One of "json" or "xml"}

\item{\code{verbose}}{logical set to TRUE for copious printed output}

\item{\code{...}}{additional parameters passed to curl
Upload new holidays}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIDAS-upload_holiday"></a>}}
\if{latex}{\out{\hypertarget{method-MIDAS-upload_holiday}{}}}
\subsection{Method \code{upload_holiday()}}{
Requires special CEC approval. Upload access is
generally only granted to load serving entities, with rare exceptions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIDAS$upload_holiday(filename, upload_format = "json", verbose = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{path to file with XML to upload}

\item{\code{upload_format}}{character One of "json" or "xml"}

\item{\code{verbose}}{logical set to TRUE for copious printed output}

\item{\code{...}}{additional parameters passed to crul options}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIDAS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MIDAS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIDAS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
