% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_support.R
\name{rate_to_json}
\alias{rate_to_json}
\title{Convert streaming rate to JSON for upload}
\usage{
rate_to_json(DT, prettify = FALSE)
}
\arguments{
\item{DT}{data.table created by the TOU_to_streaming function}

\item{prettify}{atomic logical TRUE to generate prettified JSON, FALSE by default}
}
\description{
Function to encode the data.table output of TOU_to_streaming() or csv_to_streaming()
into JSON for upload to MIDAS.
}
