% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_support.R
\name{rate_to_xml}
\alias{rate_to_xml}
\title{Convert streaming rate to XML for upload}
\usage{
rate_to_xml(DT)
}
\arguments{
\item{DT}{data.table created by the TOU_to_streaming function}
}
\value{
An atomic character containing the rate data encoded as xml
}
\description{
Function to encode the data.table output of the TOU_to_streaming function
into XML for upload to MIDAS.
}
\examples{
holidays <- import_holidays(holiday_file)
DT <- TOU_to_streaming(holidays, tou_file, start_date, end_date)

rate_to_xml(DT)
}
