% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_support.R
\name{save_rate}
\alias{save_rate}
\title{Save JSON or XML encoded rate or JSON encoded holiday to a file}
\usage{
save_rate(encoded_rate, file_name)
}
\arguments{
\item{encoded_rate}{atomic character created by rate_to_xml or rate_to_json}

\item{file_name}{atomic character with the path where you want to save the file}
}
\description{
Function to save JSON or XML encoded rate to a file.
This also works to save encoded holidays to a file.
If the file already exists, it will be overwritten with a warning.
}
