% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_support.R
\name{TOU_to_streaming}
\alias{TOU_to_streaming}
\title{TOU to streaming rate converter}
\usage{
TOU_to_streaming(
  holidays,
  tou_file,
  start_date,
  end_date,
  time_zone = "America/Los_Angeles"
)
}
\arguments{
\item{holidays}{data.frame of holiday dates and information from import_holidays()}

\item{tou_file}{atomic character with the path to a CSV file with TOU rate information}

\item{start_date}{atomic character or Date for the beginning of the period the streaming rate should cover}

\item{end_date}{atomic character or Date for the end of the period the streaming rate should cover}

\item{time_zone}{atomic character the time zone where the rate is offered. This should be one of the time zones listed in OlsonNames(). Dates and times in the CSV are assumed to be in the specified time zone.}
}
\description{
Function to import CSV files that contain TOU rate information and convert
the TOU structure to a streaming structure.
}
