% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_support.R
\name{xml_non_na}
\alias{xml_non_na}
\title{Only write non-missing XML blocks}
\usage{
xml_non_na(value, name)
}
\arguments{
\item{value}{A value to be evaluated for missingness}

\item{name}{A character string to be used as the xml tag}
}
\value{
A character string of xml
}
\description{
Returns a string of xml containing the name and value when value is non-missing
}
