% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOU to streaming.R
\name{rate_to_xml}
\alias{rate_to_xml}
\title{Convert streaming rate to XML for upload}
\usage{
rate_to_xml(DT, file_name)
}
\arguments{
\item{DT}{data.table created by the TOU_to_streaming function}

\item{file_name}{atomic character with the path where you want to save the XML}
}
\description{
Function to convert the data.table output of the TOU_to_streaming function
into XML for upload to MIDAS.
}
